Move all code inside ruby.rb crazy fun into its own namespace to make it easier to triage
