+++ rake_tasks/crazy_fun/mappings/ruby_mappings/add_test_defaults.rb
@@ -1,0 +1,10 @@
+class RubyMappings
+  class AddTestDefaults
+    def handle(_fun, dir, args)
+      args[:include] = Array(args[:include])
+      args[:include] << "#{dir}/spec"
+
+      args[:command] = args[:command] || "rspec"
+    end
+  end
+end
