+++ rake_tasks/crazy_fun/mappings/ruby_mappings/add_test_dependencies.rb
@@ -1,0 +1,11 @@
+class RubyMappings
+  class AddTestDependencies < Tasks
+    def handle(_fun, dir, args)
+      task = Rake::Task[task_name(dir, "#{args[:name]}-test")]
+
+      if args.has_key?(:deps)
+        add_dependencies task, dir, args[:deps]
+      end
+    end
+  end
+end
