+++ rake_tasks/crazy_fun/mappings/ruby_mappings/check_test_args.rb
@@ -1,0 +1,8 @@
+class RubyMappings
+  class CheckTestArgs
+    def handle(_fun, dir, args)
+      raise "no :srcs specified for #{dir}" unless args.has_key? :srcs
+      raise "no :name specified for #{dir}" unless args.has_key? :name
+    end
+  end
+end
